Rename default storage config helper
